--- a/lib/index.js
+++ b/lib/index.js
@@ -33,9 +33,9 @@ module.exports = function (levels, options) {
     function scan(level, callback) {
         var step = level.scheme(level.path, level.naming);
 
-        walk(level.path, null, callback);
+        walk(level.path, 1, null, callback);
 
-        function walk(dir, data, callback) {
+        function walk(dir, depth, data, callback) {
             fs.readdir(dir, function (err, items) {
                 if (err) return callback(err);
 
@@ -50,12 +50,13 @@ module.exports = function (levels, options) {
                         basename: basename,
                         dirname: dir,
                         filename: filename,
+                        depth: depth,
                         data: data
                     }, add, deepWalk, next);
                 }
 
                 function deepWalk(dir, data) {
-                    walk(dir, data, next);
+                    walk(dir, depth + 1, data, next);
                 }
 
                 function next(err) {