--- a/lib/index.js
+++ b/lib/index.js
@@ -33,9 +33,9 @@ module.exports = function (levels, options) {
     function scan(level, callback) {
         var step = level.scheme(level.path, level.naming);
 
-        walk(level.path, callback);
+        walk(level.path, null, callback);
 
-        function walk(dir, callback) {
+        function walk(dir, data, callback) {
             fs.readdir(dir, function (err, items) {
                 if (err) return callback(err);
 
@@ -49,12 +49,13 @@ module.exports = function (levels, options) {
                     step({
                         basename: basename,
                         dirname: dir,
-                        filename: filename
+                        filename: filename,
+                        data: data
                     }, add, deepWalk, next);
                 }
 
-                function deepWalk(dir) {
-                    walk(dir, next);
+                function deepWalk(dir, data) {
+                    walk(dir, data, next);
                 }
 
                 function next(err) {