--- a/lib/schemes/nested.js
+++ b/lib/schemes/nested.js
@@ -1,11 +1,10 @@
-var path = require('path'),
-    sep = path.sep;
+var path = require('path');
 
 module.exports = function (level, naming) {
     return function (item, add, walk, next) {
         var basename = item.basename,
             filename = item.filename,
-            depth = filename.split(sep).length - level.split(sep).length,
+            depth = item.depth,
             splited = basename.split('.'),
             dirNotation, dirType;