--- a/lib/schemes/nested.js
+++ b/lib/schemes/nested.js
@@ -5,29 +5,66 @@ module.exports = function (level, naming) {
     return function (item, add, walk, next) {
         var basename = item.basename,
             filename = item.filename,
-            parent = path.basename(item.dirname),
-            grandParent = path.basename(path.dirname(item.dirname)),
-            depth = filename.split(sep).length - level.split(sep).length;
+            depth = filename.split(sep).length - level.split(sep).length,
+            splited = basename.split('.'),
+            dirNotation, dirType;
 
-        if ((depth === 1 && naming.isBlock(basename)) ||
-            (depth === 2 && naming.isBlockMod(parent + basename)) ||
-            (depth === 2 && naming.isElem(parent + basename)) ||
-            (depth === 3 && naming.isElemMod(grandParent + parent + basename))
-        ) {
-            return walk(filename);
-        }
+        if (splited.length === 1) {
+            var dirname = item.dirname,
+                parent;
 
-        var splited = basename.split('.'),
-            bemname = splited.shift(),
-            notation = naming.parse(bemname),
-            tech =  splited.join('.');
+            if (depth === 1) {
+                dirNotation = naming.parse(basename);
+                dirType = naming.typeOf(dirNotation);
 
-        if (notation && tech) {
-            notation.level = level;
-            notation.tech = tech;
-            notation.path = filename;
+                if (dirType === 'block') {
+                    return walk(filename, { notation: dirNotation, type: dirType });
+                }
+            } else if (depth === 2) {
+                parent = path.basename(dirname);
+                dirNotation = naming.parse(parent + basename);
+                dirType = naming.typeOf(dirNotation);
 
-            add(notation);
+                if (dirType === 'blockMod' || dirType === 'elem') {
+                    return walk(filename, { notation: dirNotation, type: dirType });
+                }
+            } else if (depth === 3) {
+                var grandParent = path.basename(path.dirname(dirname));
+
+                parent = path.basename(dirname);
+                dirNotation = naming.parse(grandParent + parent + basename);
+                dirType = naming.typeOf(dirNotation);
+
+                if (dirType === 'elemMod') {
+                    return walk(filename, { notation: dirNotation, type: dirType });
+                }
+            }
+        } else {
+            if (depth > 1) {
+                var data = item.data,
+                    bemname = splited.shift(),
+                    tech = splited.join('.'),
+                    notation = naming.parse(bemname);
+
+                dirNotation = data.notation;
+                dirType = data.type;
+
+                if (notation &&
+                    (depth === 2 && dirNotation.block === notation.block) ||
+                    (depth === 3 && dirType === 'elem' && dirNotation.block === notation.block &&
+                        dirNotation.elem === notation.elem) ||
+                    (depth === 3 && dirType === 'blockMod' && dirNotation.block === notation.block &&
+                        dirNotation.modName === notation.modName) ||
+                    (depth === 4 && dirNotation.block === notation.block && dirNotation.elem === notation.elem &&
+                        dirNotation.modName === notation.modName)
+                ) {
+                    notation.level = level;
+                    notation.tech = tech;
+                    notation.path = filename;
+
+                    add(notation);
+                }
+            }
         }
 
         next();