* Fixed scheme: ignore some files if name of BEM-entity not match with parent dirnames (see tests).
* Accelerated the walk.

### Benchmarks

before:

```
           1,132 op/s » nested level
              53 op/s » `bem-bl`
              32 op/s » `bem-core` + `bem-components`
              18 op/s » `bem-bl` + `bem-core` + `bem-components`
```

after: 
```
            bem-walk
           2,717 op/s » nested level
             147 op/s » `bem-bl`
              99 op/s » `bem-core` + `bem-components`
              59 op/s » `bem-bl` + `bem-core` + `bem-components`
```

Third-party:
```
                      enb
           2,662 op/s » nested level
              71 op/s » `bem-bl`
              50 op/s » `bem-core` + `bem-components`
              29 op/s » `bem-bl` + `bem-core` + `bem-components`

                      scan-level
           2,972 op/s » nested level
              77 op/s » `bem-bl`
              52 op/s » `bem-core` + `bem-components`
              30 op/s » `bem-bl` + `bem-core` + `bem-components`
```