@tadatuta 2015-01-04
@zxqfox looks like @andrewblond  just found the way to make `bem-walk` 9 times faster! ;)