@andrewblond 2015-01-04
> Ok. I'm going to use scan-level ;-D

`bem-walk` is faster for real libs ;) `nested level` is fixtures: [benchmark/fixtures/nested-level](benchmark/fixtures/nested-level).