@zxqfox 2015-01-04
You should try `node --prof tests` with node v0.11, and something like `node-tick-processor` to analyze `v8.log`. Very simple and useful